Restructure Folders
With the growth of this project, the folders needed
to be restructured for better oversight.
