+++ includes/platforms/darwin/macUtils.hpp
