+++ includes/platforms/linux/posixUtils.hpp
