+++ includes/platforms/windows/winUtils.hpp
