+++ includes/qst/platforms.hpp
