+++ includes/qst/syncconnector.h
