+++ resources/Info.plist
