+++ resources/Syncthing.icns
