+++ resources/images/Syncthing.icns
