+++ resources/images/makeIcon.sh
