+++ resources/images/syncthingBlack.png
