+++ resources/images/syncthingBlackAnim.gif
