+++ resources/images/syncthingBlue.png
