+++ resources/images/syncthingBlueAnim.gif
