+++ resources/qsyncthing.qrc
