+++ scripts/clean.sh
