+++ scripts/generate_msvc.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+# For OS X
+
+./clean.sh
+/c/Qt/5.5/msvc2013_64/bin/qmake -tp vc ../QSyncthingTray.pro
