+++ scripts/generate_xcode.sh
@@ -2,5 +2,5 @@
 # For OS X
 
 ./clean.sh
-~/Qt/5.5/clang_64/bin/qmake -spec macx-xcode .
+~/Qt/5.5/clang_64/bin/qmake -spec macx-xcode ../
 open QSyncthingTray.xcodeproj
