+++ sources/qst/syncwebpage.cpp
@@ -17,7 +17,7 @@
  ******************************************************************************/
 
 #include <iostream>
-#include "syncwebpage.h"
+#include <qst/syncwebpage.h>
 
 //------------------------------------------------------------------------------------//
 #define UNUSED(x) (void)(x)
