--- src/CMakeLists.txt
@@ -1,11 +1,0 @@
-##  --------------------------------------------------------------------------------------
-##   Create Executable
-##  --------------------------------------------------------------------------------------
-
-
-set(SRC_ROOT ${CMAKE_CURRENT_SOURCE_DIR})
-file(GLOB_RECURSE FILES_WP RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} ${SRC_ROOT}/*.cpp ${SRC_ROOT}/*.h(pp))
-
-##  --------------------------------------------------------------------------------------
-##   Dependencies
-##  --------------------------------------------------------------------------------------
